PK 타입 변경

+++ src/main/java/sshj/sshj/mapper/ClubMapper.java
@@ -11,23 +11,23 @@ import java.util.List;
 @Mapper
 public interface ClubMapper {
     void insertClubNotice(ClubNoticeDto dto) throws Exception;
-    ClubNoticeDto selectClubNotice(int noticeId) throws Exception;
-    List<ClubNoticeDto> selectClubNoticeList(int clubId) throws Exception;
+    ClubNoticeDto selectClubNotice(long noticeId) throws Exception;
+    List<ClubNoticeDto> selectClubNoticeList(long clubId) throws Exception;
     void updateClubNotice(ClubNoticeDto dto) throws Exception;
-    void deleteClubNotice(int noticeId) throws Exception;
-    void insertClubNoticeLike(int userId,int noticeId) throws Exception;
-    void deleteClubNoticeLike(int userId,int noticeId) throws Exception;
-    int selectClubNoticeCnt(int noticeId) throws Exception;
-    void insertClubSubs(int userId,int clubId) throws Exception;
-    void deleteClubSubs(int userId,int clubId) throws Exception;
-    List selectClubSubsUserList(int clubId) throws Exception;
+    void deleteClubNotice(long noticeId) throws Exception;
+    void insertClubNoticeLike(long userId,long noticeId) throws Exception;
+    void deleteClubNoticeLike(long userId,long noticeId) throws Exception;
+    int selectClubNoticeCnt(long noticeId) throws Exception;
+    void insertClubSubs(long userId,long clubId) throws Exception;
+    void deleteClubSubs(long userId,long clubId) throws Exception;
+    List selectClubSubsUserList(long clubId) throws Exception;
 
 
     void insertClubDescription(ClubDescriptionDto clubDescriptionDto) throws Exception;
-    String selectClubDescription(int clubId) throws Exception;
+    String selectClubDescription(long clubId) throws Exception;
     void updateClubDescription(ClubDescriptionDto clubDescriptionDto) throws Exception;
-    void deleteClubDescription(int clubId) throws Exception;
+    void deleteClubDescription(long clubId) throws Exception;
 
-    int selectIsSubClub(int userId, int clubId) throws Exception;
-    List<Integer> selectSubClubList(int userId) throws Exception;
+    int selectIsSubClub(long userId, long clubId) throws Exception;
+    List<Long> selectSubClubList(long userId) throws Exception;
 }
